refactor: move reevaluateSeedSpecies to core engine and import into parameterScan handler
Moves the handler-local reevaluateSeedSpecies logic into the shared engine package under utils/paramUtils.ts and exposes it via src/index.ts to ensure strict engine/MCP handler parity.
