+++ packages/mcp-server/src/handlers/parameterScan.ts
@@ -1,29 +1,8 @@
-import { generateRange, simulate, loadEvaluator, BNGLParser } from '@bngplayground/engine';
+import { generateRange, simulate, loadEvaluator, reevaluateSeedSpecies } from '@bngplayground/engine';
 import { ToolArgs, ToolResult } from '../types/index.js';
 import { parameterScanArgsSchema } from '../schemas/index.js';
 import { createToolResult, parseArgs, applyNetworkOptions, parseModelOrThrow, buildSimulationOptions, expandModel, assertScannableParameter, cloneExpandedModel, updateMassActionRates } from '../services/engine.js';
 import { structureError } from '../services/errors.js';
-
-function reevaluateSeedSpecies(model: any, seedExpressions: Map<string, string>): void {
-  const paramMap = new Map<string, number>(Object.entries(model.parameters ?? {}));
-  const functionMap = new Map<string, { args: string[]; expr: string }>(
-    (model.functions ?? []).map((fn: any) => [fn.name, { args: fn.args ?? [], expr: fn.expression ?? '' }]),
-  );
-
-  for (const species of model.species ?? []) {
-    const fallbackExpression = seedExpressions.get(species.name);
-    const expr = typeof species.initialExpression === 'string'
-      ? species.initialExpression.trim()
-      : typeof fallbackExpression === 'string'
-        ? fallbackExpression.trim()
-        : '';
-    if (!expr) continue;
-    const evaluated = BNGLParser.evaluateExpression(expr, paramMap, undefined, functionMap);
-    if (Number.isFinite(evaluated)) {
-      species.initialConcentration = evaluated;
-    }
-  }
-}
 
 export async function handleParameterScan(args: ToolArgs): Promise<ToolResult<any>> {
   try {
